+++ python/bullmq/event_emitter.py
@@ -7,12 +7,14 @@ class EventEmitter:
     def __init__(self):
         self._callbacks: Dict[str, callable] = {}
 
-    def on(self, event_name, function):
-        self._callbacks[event_name] = self._callbacks.get(event_name, []) + [function]
+    def on(self, event_name: str, function):
+        self._callbacks[event_name] = self._callbacks.get(
+            event_name, []) + [function]
         return function
 
-    def emit(self, event_name, *args, **kwargs):
-        [function(*args, **kwargs) for function in self._callbacks.get(event_name, [])]
+    def emit(self, event_name: str, *args, **kwargs):
+        [function(*args, **kwargs)
+         for function in self._callbacks.get(event_name, [])]
 
-    def off(self, event_name, function):
+    def off(self, event_name: str, function):
         self._callbacks.get(event_name, []).remove(function)