+++ python/bullmq/worker.py
@@ -1,20 +1,73 @@
-from typing import Callable
+from typing import Callable, TypedDict, Any
 from uuid import uuid4
-import asyncio
-import traceback
-import time
-
-from redis import Redis
-
 from bullmq.scripts import Scripts
 from bullmq.redis_connection import RedisConnection
 from bullmq.event_emitter import EventEmitter
 from bullmq.job import Job
 from bullmq.timer import Timer
 
+import asyncio
+import traceback
+import time
+
+
+class WorkerOptions(TypedDict, total=False):
+    autorun: bool
+    """
+    Condition to start processor at instance creation
+
+    @default true 
+    """
+
+    concurrency: int
+    """
+    Amount of jobs that a single worker is allowed to work on
+    in parallel.
+   
+    @default 1
+    @see https://docs.bullmq.io/guide/workers/concurrency
+    """
+
+    maxStalledCount: int
+    """
+    Amount of times a job can be recovered from a stalled state
+    to the `wait` state. If this is exceeded, the job is moved
+    to `failed`.
+   
+    @default 1
+    """
+
+    stalledInterval: int
+    """
+    Number of milliseconds between stallness checks.
+   
+    @default 30000
+    """
+
+    lockDuration: int
+    """
+    Duration of the lock for the job in milliseconds. The lock represents that
+    a worker is processing the job. If the lock is lost, the job will be eventually
+    be picked up by the stalled checker and move back to wait so that another worker
+    can process it again.
+   
+    @default 30000 
+    """
+
+    prefix: str
+    """
+    Prefix for all queue keys.
+    """
+
+    connection: dict[str, Any]
+    """
+    Options for connecting to a Redis instance.
+    """
+
+
 class Worker(EventEmitter):
-    def __init__(self, name: str, processor: Callable[[Job, str], asyncio.Future], opts = {}):
-        super().__init__()  
+    def __init__(self, name: str, processor: Callable[[Job, str], asyncio.Future], opts: WorkerOptions = {}):
+        super().__init__()
         self.name = name
         self.processor = processor
         self.opts = {
@@ -23,17 +76,12 @@ class Worker(EventEmitter):
             "maxStalledCount": opts.get("maxStalledCount", 1),
             "stalledInterval": opts.get("stalledInterval", 30000),
         }
-
-        redisOpts = opts.get("connection") or {}
-
-        self.redisConnection = RedisConnection(redisOpts)
+        redis_opts = opts.get("connection", {})
+        self.redisConnection = RedisConnection(redis_opts)
+        self.blockingRedisConnection = RedisConnection(redis_opts)
         self.client = self.redisConnection.conn
-
-        self.blockingRedisConnection = RedisConnection(redisOpts)
         self.bclient = self.blockingRedisConnection.conn
-    
-        self.scripts = Scripts(opts.get("prefix") or "bull", name, self.client)
-
+        self.scripts = Scripts(opts.get("prefix", "bull"), name, self.client)
         self.closing = False
         self.forceClosing = False
         self.closed = False
@@ -48,26 +96,29 @@ class Worker(EventEmitter):
         if self.running:
             raise Exception("Worker is already running")
 
-        self.timer = Timer((self.opts.get("lockDuration") / 2) / 1000, self.extendLocks)
-        self.stalledCheckTimer = Timer(self.opts.get("stalledInterval") / 1000, self.runStalledJobsCheck)
+        self.timer = Timer(
+            (self.opts.get("lockDuration") / 2) / 1000, self.extendLocks)
+        self.stalledCheckTimer = Timer(self.opts.get(
+            "stalledInterval") / 1000, self.runStalledJobsCheck)
         self.running = True
         job = None
-
         token = uuid4().hex
+
         while not self.closed:
             if not job and len(self.processing) < self.opts.get("concurrency") and not self.closing:
-                waitingJob = asyncio.ensure_future(self.getNextJob(token))
-                self.processing.add(waitingJob)
+                waiting_job = asyncio.ensure_future(self.getNextJob(token))
+                self.processing.add(waiting_job)
 
             if job:
-                processingJob = asyncio.ensure_future(self.processJob(job, token))
-                self.processing.add(processingJob)
+                processing_job = asyncio.ensure_future(
+                    self.processJob(job, token))
+                self.processing.add(processing_job)
 
             try:
                 job, pending = await getFirstCompleted(self.processing)
                 self.processing = pending
 
-                if job == None or len(self.processing) == 0 and self.closing:
+                if (job is None or len(self.processing) == 0) and self.closing:
                     # We are done processing so we can close the queue
                     break
 
@@ -82,23 +133,29 @@ class Worker(EventEmitter):
         self.stalledCheckTimer.stop()
 
     async def getNextJob(self, token: str):
+        """
+        Returns a promise that resolves to the next job in queue.
+        @param token: worker token to be assigned to retrieved job
+        @returns a Job or undefined if no job was available in the queue.
+        """
         # First try to move a job from the waiting list to the active list
         result = await self.scripts.moveToActive(token, self.opts)
         job = None
-        jobId = None
-        delayUntil = None
+        job_id = None
+        delay_until = None
         if result:
-            job, jobId = result
+            job, job_id = result
 
         # If there are no jobs in the waiting list we keep waiting with BRPOPLPUSH
-        if job == None:
-            timeout = min(delayUntil - int(time.time() * 1000) if delayUntil else 5000, 5000) / 1000
-            jobId = await self.bclient.brpoplpush(self.scripts.keys["wait"], self.scripts.keys["active"], timeout)
-            if jobId:
-                job, jobId = await self.scripts.moveToActive(token, self.opts, jobId)
+        if job is None:
+            timeout = min(delay_until - int(time.time() * 1000)
+                          if delay_until else 5000, 5000) / 1000
+            job_id = await self.bclient.brpoplpush(self.scripts.keys["wait"], self.scripts.keys["active"], timeout)
+            if job_id:
+                job, job_id = await self.scripts.moveToActive(token, self.opts, job_id)
 
-        if job and jobId:
-            return Job.fromJSON(self.client, job, jobId)
+        if job and job_id:
+            return Job.fromJSON(self.client, job, job_id)
 
     async def processJob(self, job: Job, token: str):
         try:
@@ -116,7 +173,7 @@ class Worker(EventEmitter):
                     await self.scripts.moveToFailed(job, str(err), job.opts.get("removeOnFail", False), token, self.opts, fetchNext=not self.closing)
 
                 # TODO: Store the stacktrace in the job
-                
+
                 self.emit("failed", job, err)
             except Exception as err:
                 print("Error moving job to failed", err)
@@ -136,7 +193,7 @@ class Worker(EventEmitter):
             # We should emit an error for each of those jobs.
             #    for jobId, err in result.items():
             #    self.emit("error", "could not renew lock for job " + jobId)
-            
+
         except Exception as e:
             print("Error renewing locks", e)
             traceback.print_exc()
@@ -145,7 +202,8 @@ class Worker(EventEmitter):
         try:
             failed, stalled = await self.scripts.moveStalledJobsToWait(self.opts.get("maxStalledCount"), self.opts.get("stalledInterval"))
             for jobId in failed:
-                self.emit("failed", jobId, "job stalled more than allowable limit")
+                self.emit("failed", jobId,
+                          "job stalled more than allowable limit")
             for jobId in stalled:
                 self.emit("stalled", jobId)
 
@@ -154,7 +212,9 @@ class Worker(EventEmitter):
             self.emit('error', e)
 
     async def close(self, force: bool = False):
-        """ "Close the worker" """
+        """
+        Close the worker
+        """
         if force:
             self.forceClosing = True
 
@@ -163,10 +223,11 @@ class Worker(EventEmitter):
         await self.blockingRedisConnection.close()
         await self.redisConnection.close()
 
-async def getFirstCompleted( taskSet: set):
+
+async def getFirstCompleted(taskSet: set):
     jobSet, pending = await asyncio.wait(taskSet, return_when=asyncio.FIRST_COMPLETED)
     for jobTask in jobSet:
-        try: 
+        try:
             job = jobTask.result()
             return (job, pending)
         except Exception as e: